chore: moved bulk operations to Vizpay Transactions py file
