+++ vizpay/hooks.py
@@ -168,7 +168,7 @@ app_license = "agpl-3.0"
 
 scheduler_events = {
 	"hourly_long": [
-		"vizpay.bulk_transactions.fetch_status_for_pending"
+		"vizpay.vizpay.doctype.vizpay_transaction.vizpay_transaction.fetch_status_for_pending"
 	],
 }
